# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Molecular Vision AR (MVAR) API Documentation",
        "description": "Dokumentasi RESTful API untuk Molecular Vision AR (MVAR) Backend. API ini digunakan oleh aplikasi mobile (Flutter / Unity) dan client lainnya.",
        "contact": {
            "name": "Tim Pengembang MVAR",
            "email": "dev@mvar.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Current API Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/molecules": {
            "get": {
                "tags": [
                    "Molekul"
                ],
                "summary": "Daftar Katalog Molekul 3D",
                "description": "Mengambil daftar seluruh molekul kimia 3D dengan opsi pencarian (berdasarkan nama atau rumus kimia) dan filter bentuk geometri.",
                "operationId": "e554a52fa5b17bfb6bdda5a5ed4c043c",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci pencarian nama atau rumus molekul (misal: Air, H2O, CH4)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "H2O"
                        }
                    },
                    {
                        "name": "shape",
                        "in": "query",
                        "description": "Filter berdasarkan bentuk geometri molekul (misal: Linear, Bent, Tetrahedral)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Bent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar molekul berhasil diambil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar molekul berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_molecule": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Air"
                                                    },
                                                    "formula": {
                                                        "type": "string",
                                                        "example": "H2O"
                                                    },
                                                    "shape": {
                                                        "type": "string",
                                                        "example": "Bengkok (Bent)"
                                                    },
                                                    "bent": {
                                                        "type": "string",
                                                        "example": "104.5°"
                                                    },
                                                    "bond_type": {
                                                        "type": "string",
                                                        "example": "Kovalen Polar"
                                                    },
                                                    "model_3d_url": {
                                                        "type": "string",
                                                        "example": "https://storage.googleapis.com/.../h2o.glb"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - Token tidak valid atau tidak disertakan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/molecules/{id}": {
            "get": {
                "tags": [
                    "Molekul"
                ],
                "summary": "Detail Molekul 3D",
                "description": "Mengambil detail lengkap suatu molekul kimia termasuk deskripsi teori dan URL model 3D.",
                "operationId": "204af82e0f4082a9ea42fe7a8e611b53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID molekul kimia",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail molekul berhasil diambil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail molekul berhasil diambil."
                                        },
                                        "data": {
                                            "properties": {
                                                "id_molecule": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Air"
                                                },
                                                "formula": {
                                                    "type": "string",
                                                    "example": "H2O"
                                                },
                                                "shape": {
                                                    "type": "string",
                                                    "example": "Bengkok (Bent)"
                                                },
                                                "bent": {
                                                    "type": "string",
                                                    "example": "104.5°"
                                                },
                                                "bond_type": {
                                                    "type": "string",
                                                    "example": "Kovalen Polar"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Molekul air terdiri atas dua atom hidrogen yang berikatan kovalen dengan satu atom oksigen..."
                                                },
                                                "model_3d_url": {
                                                    "type": "string",
                                                    "example": "https://storage.googleapis.com/.../h2o.glb"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - Token tidak valid atau tidak disertakan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Molekul kimia tidak ditemukan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Molekul kimia tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan token Sanctum dengan format: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}